Remove NativeERC20 and rename Mintable
